TANKX/analyse/requirements.txt
redis
numpy
python-binance